overlapScore

+++ overlapScore.py
@@ -67,7 +67,7 @@ def getProbQuality (q):
     #p = 10**(-np.float128(q)/10)
     #p = 10 ** (-np.float128(qNew)/10)
     #print(p)
-    return qNew
+    return np.float128(qNew)
 
 """
 @Definition:
